requirements.txt
spotipy
click
django
django-rest-framework
jsonfield
django-filter
djangorestframework-filters